git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@1304 01821946-851b-0410-b7f4-f1477c70e57a

+++ network/network_phandler.cpp
@@ -139,8 +139,8 @@ if (gameStates.multi.nGameType == UDP_GAME)
 	 addressFilter [PID_EXTRA_GAMEINFO] =
 	 addressFilter [PID_DOWNLOAD] =
 	 addressFilter [PID_UPLOAD] =
-	 addressFilter [PID_TRACKER_ADD_SERVER] =
-	 addressFilter [PID_TRACKER_GET_SERVERLIST] = 1;
+	 addressFilter [(int) PID_TRACKER_ADD_SERVER] =
+	 addressFilter [(int) PID_TRACKER_GET_SERVERLIST] = 1;
 }
 
 //------------------------------------------------------------------------------
